requirements.txt
aiohttp==3.8.5
gspread==5.11.3
google-auth-oauthlib==1.1.0
google-auth==2.25.2
beautifulsoup4==4.12.2
requests==2.31.0
python-telegram-bot==20.3
